requirements.txt
streamlit==1.35.0
pillow==10.3.0